+++ tests/testthat/test_user_input.R
@@ -1,4 +1,3 @@
-
 ## Function to suppress print/cat outputs
 quiet <- function(x) {
   sink(tempfile())
@@ -7,7 +6,6 @@ quiet <- function(x) {
 }
 
 test_that("User Input", {
-
   area <- sf::st_as_sf(sf::st_sfc(
     sf::st_polygon(list(cbind(
       c(4498482, 4498482, 4499991, 4499991, 4498482),
@@ -29,126 +27,32 @@ test_that("User Input", {
     "selection_mode"
   )
 
+  polygon <- area
+  id <- resultrect[1, "bestPaEn"][[1]][1, "Rect_ID"]
+  new <- random_search_single(resultrect, polygon, turbine = id, runs = 5)
+  expect_type(new, "list")
+  expect_false(anyNA(unlist(new)))
+  new_df <- do.call(rbind, new)
+  expect_true(all(new_df[, "EfficAllDir"] <= 100 & new_df[, "EfficAllDir"] > 0))
+  expect_true(all(new_df[, "EnergyOverall"] > 0))
+  expect_true(all(new_df[, "AbschGesamt"] >= 0))
 
-  ## random_search_single - Rects #############
-  polygon <- sf::st_as_sf(sf::st_sfc(
-    sf::st_polygon(list(cbind(
-      c(4498482, 4498482, 4499991, 4499991, 4498482),
-      c(2668272, 2669343, 2669343, 2668272, 2668272)
-    ))),
-    crs = 3035
+  new <- quiet(random_search_single(
+    resultrect, polygon, turbine = id, runs = 3,
+    max_dist = 5, plot = TRUE
   ))
-
-  ## which IDs
-  resldat <- do.call("rbind", resultrect[, "bestPaEn"])
-  maxDist <- as.numeric(resultrect[, "inputData"][[1]]["Rotorradius", ]) * 2.2
-  resldat <- as.data.frame(resldat[!duplicated(resldat[, "Run"]), , drop = FALSE])
-  resldat$GARun <- 1:nrow(resldat)
-  resldat <- resldat[order(resldat[, 4], decreasing = TRUE), ]
-  bestGARun <- resldat$GARun[1]
-  resolu <- as.numeric(resultrect[bestGARun, ]$inputData["Resolution", ][1])
-  rotRad <- as.numeric(resultrect[bestGARun, ]$inputData["Rotorradius", ][1])
-  propu <- as.numeric(resultrect[bestGARun, ]$inputData["Percentage of Polygon", ][1])
-  winddata <- resultrect[bestGARun, ]$inputWind
-  Grid <- grid_area(
-    area = polygon, size = resolu,
-    prop = propu, plot_grid = FALSE
+  expect_type(new, "list")
+  plres <- plot_random_search(
+    resultRS = new, result = resultrect, area = polygon, best = 2
   )
-  maxFac <- rotRad * (resolu / (rotRad * 2))
-  winddata <- windata_format(winddata)
-  probabDir <- winddata[[2]]
-  winddata <- winddata[[1]]
-  layout_start <- resultrect[bestGARun, ]$bestPaEn
-
-  id <- sample(layout_start[, "Rect_ID"], size = 100, TRUE)
-  f <- file()
-  options(windfarmGA.connection = f)
-  ans <- paste(id, collapse = "\n")
-  write(ans, f)
-
-  new <- random_search_single(resultrect, polygon)
-  expect_type(new, "list")
-  expect_false(anyNA(unlist(new)))
-  new_df <- do.call(rbind, new)
-  expect_true(all(new_df[, "EfficAllDir"] <= 100 & new_df[, "EfficAllDir"] > 0))
-  expect_true(all(new_df[, "EnergyOverall"] > 0))
-  expect_true(all(new_df[, "AbschGesamt"] >= 0))
-  rm(new, new_df)
-
-  for (i in 1:40) {
-    new <- quiet(random_search_single(resultrect, polygon, max_dist = 5, plot = TRUE))
-    expect_type(new, "list")
-    expect_false(anyNA(unlist(new)))
-    new_df <- do.call(rbind, new)
-    expect_true(all(new_df[, "EfficAllDir"] <= 100 & new_df[, "EfficAllDir"] > 0))
-    expect_true(all(new_df[, "EnergyOverall"] > 0))
-    expect_true(all(new_df[, "AbschGesamt"] >= 0))
-    rm(new, new_df)
-  }
-
-  new <- random_search_single(resultrect, polygon)
-  expect_type(new, "list")
-  expect_false(anyNA(unlist(new)))
-  new_df <- do.call(rbind, new)
-  expect_true(all(new_df[, "EfficAllDir"] <= 100 & new_df[, "EfficAllDir"] > 0))
-  expect_true(all(new_df[, "EnergyOverall"] > 0))
-  expect_true(all(new_df[, "AbschGesamt"] >= 0))
-
-  plres <- plot_random_search(resultRS = new, result = resultrect, area = polygon, best = 2)
   expect_true(is.null(plres))
 
-  # reset connection
-  options(windfarmGA.connection = stdin())
-  # close the file
-  close(f)
-
-
-  ## random_search_single - Hexagons ##################
-  polygon <- sf::st_as_sf(sf::st_sfc(
-    sf::st_polygon(list(cbind(
-      c(4498482, 4498482, 4499991, 4499991, 4498482),
-      c(2668272, 2669343, 2669343, 2668272, 2668272)
-    ))),
-    crs = 3035
-  ))
-
-  ## which IDs
-  resldat <- do.call("rbind", resulthex[, "bestPaEn"])
-  resldat <- as.data.frame(resldat[!duplicated(resldat[, "Run"]), , drop = FALSE])
-  resldat$GARun <- 1:nrow(resldat)
-  resldat <- resldat[order(resldat[, 4], decreasing = TRUE), ]
-  bestGARun <- resldat$GARun[1]
-  resolu <- as.numeric(resulthex[bestGARun, ]$inputData["Resolution", ][1])
-  rotRad <- as.numeric(resulthex[bestGARun, ]$inputData["Rotorradius", ][1])
-  maxDist <- as.numeric(resulthex[, "inputData"][[1]]["Rotorradius", ]) * 2.2
-  propu <- as.numeric(resulthex[bestGARun, ]$inputData["Percentage of Polygon", ][1])
-  winddata <- resulthex[bestGARun, ]$inputWind
-  Grid <- hexa_area(polygon, size = resolu)
-  maxFac <- rotRad * (resolu / (rotRad * 2))
-  winddata <- windata_format(winddata)
-  probabDir <- winddata[[2]]
-  winddata <- winddata[[1]]
-  layout_start <- resulthex[bestGARun, ]$bestPaEn
-
-  id <- sample(layout_start[, "Rect_ID"], 2)
-  f <- file()
-  options(windfarmGA.connection = f)
-  ans <- paste(id, collapse = "\n")
-  write(ans, f)
-
-  new <- random_search_single(resulthex, polygon)
+  id_hex <- resulthex[1, "bestPaEn"][[1]][1, "Rect_ID"]
+  new <- random_search_single(resulthex, polygon, turbine = id_hex, runs = 5)
   expect_type(new, "list")
   expect_false(anyNA(unlist(new)))
   new_df <- do.call(rbind, new)
   expect_true(all(new_df[, "EfficAllDir"] <= 100 & new_df[, "EfficAllDir"] > 0))
   expect_true(all(new_df[, "EnergyOverall"] > 0))
   expect_true(all(new_df[, "AbschGesamt"] >= 0))
-  rm(new, new_df)
-
-  # reset connection
-  options(windfarmGA.connection = stdin())
-  # close the file
-  close(f)
-
 })
-